python-analytics/requirements.txt
watchdog==3.0.0
pandas==2.1.4
numpy==1.26.2
python-dotenv==1.0.0
redis==5.0.1
paho-mqtt==1.6.1
psycopg2-binary==2.9.9
